+++ encoding/protobuf/error_test.go
@@ -213,11 +213,18 @@ func TestConvertFromYARPCError(t *testing.T) {
 }
 
 func TestCreateStatusWithDetailErrors(t *testing.T) {
-	t.Run("unsupported code", func(t *testing.T) {
+	t.Run("code ok", func(t *testing.T) {
 		pberr := &pberror{code: yarpcerrors.CodeOK, message: "test"}
 		_, err := createStatusWithDetail(pberr, Encoding, &codec{})
 		assert.Error(t, err, "unexpected empty error")
 		assert.Equal(t, err.Error(), "no status error for error with code OK")
+	})
+
+	t.Run("unsupported code", func(t *testing.T) {
+		pberr := &pberror{code: 99, message: "test"}
+		_, err := createStatusWithDetail(pberr, Encoding, &codec{})
+		assert.Error(t, err, "unexpected empty error")
+		assert.Equal(t, err.Error(), "no status error for error with code 99")
 	})
 
 	t.Run("unsupported encoding", func(t *testing.T) {
@@ -235,6 +242,12 @@ func TestErrorHandling(t *testing.T) {
 	})
 	t.Run("GetErrorDetail non pberror", func(t *testing.T) {
 		assert.Nil(t, GetErrorDetails(errors.New("test")), "unexpected details")
+	})
+	t.Run("GetErrorDetail with no error detail", func(t *testing.T) {
+		err := NewError(
+			yarpcerrors.CodeAborted,
+			"aborted")
+		assert.Nil(t, GetErrorDetails(err), "unexpected details")
 	})
 	t.Run("PbError empty error handling", func(t *testing.T) {
 		var pbErr *pberror