+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.54.1] - 2021-06-01
+### Fixed
+- protobuf error: a protobuf error created with an invalid code returns a unknown
+  error.
+- protobuf error details: protobuf.GetErrorDetails returns nil instead of an empty array
+  if there is no details in the given error.
+
 ## [1.54.0] - 2021-06-01
 ### Added
 - observability: extend response, and request payload size histogram buckets up to 256MB.
@@ -1357,6 +1364,7 @@ This release requires regeneration of ThriftRW code.
 
 - Initial release.
 
+[1.54.1]: https://github.com/yarpc/yarpc-go/compare/v1.54.0...v1.54.1
 [1.54.0]: https://github.com/yarpc/yarpc-go/compare/v1.53.2...v1.54.0
 [1.53.2]: https://github.com/yarpc/yarpc-go/compare/v1.53.1...v1.53.2
 [1.53.1]: https://github.com/yarpc/yarpc-go/compare/v1.51.0...v1.52.0